src/requirements.txt
numpy==1.17.4
opencv_python==4.1.2.30
Pillow
PyQt5==5.13.2
matplotlib==3.1.2